pyproject.toml
[project]
name = "redditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "praw>=7.8.1",
    "fastapi[standard]>=0.115.12",
    "cohere>=5.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
